Sort area types by hierarchy with fallback to total count
